+++ HISTORY.md
@@ -54,6 +54,9 @@
 * **Translation Updates:**
     * German
 
+### 2025-12-22
+* objects, gui: dynamically render sprite icons instead of creating bitmaps every time
+
 ### 2025-12-20
 * store: template-projects store their magnification level if it's above 100%